+++ README.md
@@ -1,9 +1,2 @@
 # Superfast Matching Engine
-Improved matching engine written in Go (Golang)
-
-
-## Installation
-
-```javascript
-go get github.com/Pantelwar/matching-engine/engine
-```
+    Improved matching engine written in Go (Golang)
