+++ internal/app/engine/order.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/Pantelwar/matching-engine/util"
+	"github.com/Pantelwar/matching-engine/internal/app/util"
 )
 
 // Order describes the struct of the order
