+++ internal/app/engine/process_limit_order.go
@@ -1,21 +1,26 @@
 package engine
 
 import (
+	"fmt"
+
 	"github.com/Pantelwar/binarytree"
-	"github.com/Pantelwar/matching-engine/util"
+
+	"github.com/Pantelwar/matching-engine/internal/app/util"
 )
 
-// ProcessMarket executes limit process
-func (ob *OrderBook) ProcessMarket(order Order) ([]*Order, *Order) {
+var decimalZero, _ = util.NewDecimalFromString("0.0")
+
+// Process executes limit process
+func (ob *OrderBook) Process(order Order) ([]*Order, *Order) {
 	if order.Type == Buy {
 		// return ob.processOrderB(order)
-		return ob.commonProcessMarket(order, ob.SellTree, ob.addBuyOrder, ob.removeSellNode)
+		return ob.commonProcess(order, ob.SellTree, ob.addBuyOrder, ob.removeSellNode)
 	}
 	// return ob.processOrderS(order)
-	return ob.commonProcessMarket(order, ob.BuyTree, ob.addSellOrder, ob.removeBuyNode)
+	return ob.commonProcess(order, ob.BuyTree, ob.addSellOrder, ob.removeBuyNode)
 }
 
-func (ob *OrderBook) commonProcessMarket(order Order, tree *binarytree.BinaryTree, add func(Order), remove func(float64) error) ([]*Order, *Order) {
+func (ob *OrderBook) commonProcess(order Order, tree *binarytree.BinaryTree, add func(Order), remove func(float64) error) ([]*Order, *Order) {
 	var maxNode *binarytree.BinaryNode
 	if order.Type == Sell {
 		maxNode = tree.Max()
@@ -23,9 +28,11 @@ func (ob *OrderBook) commonProcessMarket(order Order, tree *binarytree.BinaryTre
 		maxNode = tree.Min()
 	}
 	if maxNode == nil {
-		// add(order)
+		// fmt.Println("adding node pending", order.Price)
+		add(order)
 		return nil, nil
 	}
+	// fmt.Println("maxNode", maxNode.Key, maxNode.Data.(*OrderType).Tree.Root.Key)
 	count := 0
 	noMoreOrders := false
 	var allOrdersProcessed []*Order
@@ -40,14 +47,20 @@ func (ob *OrderBook) commonProcessMarket(order Order, tree *binarytree.BinaryTre
 		}
 		if maxNode == nil || noMoreOrders {
 			if order.Amount.Cmp(decimalZero) == 1 {
-				allOrdersProcessed = append(allOrdersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, decimalZero))
+				// fmt.Println("adding sell node pending")
+				add(order)
+				break
+			} else {
+				break
 			}
-			break
 		}
+		// fmt.Println("maxNode in", maxNode.Key, maxNode.Data.(*OrderType).Tree.Root.Key)
 
 		// var t []Trade
 		var ordersProcessed []*Order
-		noMoreOrders, ordersProcessed, partialOrder = ob.processLimitMarket(&order, maxNode.Data.(*OrderType).Tree, orderOriginalAmount) //, orderPrice)
+		noMoreOrders, ordersProcessed, partialOrder = ob.processLimit(&order, partialOrder, maxNode.Data.(*OrderType).Tree, orderOriginalAmount) //, orderPrice)
+		fmt.Printf("\npartialOrder in between: %#v\n", partialOrder)
+		fmt.Printf("noMoreOrders: %#v\n\n", noMoreOrders)
 		allOrdersProcessed = append(allOrdersProcessed, ordersProcessed...)
 		// trades = append(trades, t...)
 
@@ -55,16 +68,22 @@ func (ob *OrderBook) commonProcessMarket(order Order, tree *binarytree.BinaryTre
 			// node := remove(maxNode.Key)
 			// // node := ob.removeBuyNode(maxNode.Key)
 			// tree.Root = node
+			// fmt.Println("removing", maxNode.Key)
 			remove(maxNode.Key)
 		}
 	}
 
 	// return trades, allOrdersProcessed, partialOrder
+	// if partialOrder.Amount == nil {
+	// 	partialOrder = nil
+	// }
+	fmt.Printf("partialOrder final: %#v\n", partialOrder)
+
 	return allOrdersProcessed, partialOrder
 }
 
-func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTree, orderOriginalAmount *util.StandardBigDecimal) (bool, []*Order, *Order) {
-	// orderPrice, _ := order.Price.Float64()
+func (ob *OrderBook) processLimit(order, partialOrder *Order, tree *binarytree.BinaryTree, orderOriginalAmount *util.StandardBigDecimal) (bool, []*Order, *Order) {
+	orderPrice := order.Price.Float64()
 	var maxNode *binarytree.BinaryNode
 	if order.Type == Sell {
 		maxNode = tree.Max()
@@ -73,7 +92,10 @@ func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTre
 	}
 	noMoreOrders := false
 	var ordersProcessed []*Order
-	var partialOrder *Order
+	// var partialOrder *Order
+	fmt.Printf("partialOrder start: %#v\n", partialOrder)
+
+	// var partialOrder *Order
 	if maxNode == nil {
 		// return trades, noMoreOrders, nil, nil
 		return noMoreOrders, nil, nil
@@ -85,51 +107,51 @@ func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTre
 		} else {
 			maxNode = tree.Min()
 		}
+		// fmt.Println("maxNode more in", maxNode.Key, maxNode.Data.(*OrderNode).Orders)
 		if maxNode == nil || noMoreOrders {
-			break
-			// if order.Amount.Cmp(decimalZero) == 1 {
-			// 	// fmt.Println("inserting", noMoreOrders)
-			// 	// ordersProcessed = append(ordersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, decimalZero))
-			// 	// partialOrder = NewOrder(order.ID, order.Type, order.Amount, order.Price)
-			// 	break
-			// } else {
-			// 	break
-			// }
+			if order.Amount.Cmp(decimalZero) == 1 {
+				partialOrder = NewOrder(order.ID, order.Type, order.Amount, order.Price)
+				break
+			} else {
+				// partialOrder = nil
+				break
+			}
 		}
-		// if order.Type == Sell {
-		// 	if orderPrice > maxNode.Key {
-		// 		// fmt.Println("adding sellnode directly")
-		// 		noMoreOrders = true
-		// 		// return trades, noMoreOrders, nil, nil
-		// 		return noMoreOrders, nil, nil
-		// 	}
-		// } else {
-		// 	if orderPrice < maxNode.Key {
-		// 		// fmt.Println("adding buynode directly")
-		// 		noMoreOrders = true
-		// 		// return trades, noMoreOrders, nil, nil
-		// 		return noMoreOrders, nil, nil
-		// 	}
-		// }
+		if order.Type == Sell {
+			if orderPrice > maxNode.Key {
+				fmt.Println("adding sellnode directly")
+				noMoreOrders = true
+				// return trades, noMoreOrders, nil, nil
+				return noMoreOrders, ordersProcessed, partialOrder
+			}
+		} else {
+			if orderPrice < maxNode.Key {
+				fmt.Println("adding buynode directly")
+				noMoreOrders = true
+				// return trades, noMoreOrders, nil, nil
+				return noMoreOrders, ordersProcessed, partialOrder
+			}
+		}
 
 		nodeData := maxNode.Data.(*OrderNode) //([]*Order)
 		nodeOrders := nodeData.Orders         //([]*Order)
 		countMatch := 0
+		// fmt.Println("length of nodeOrders", len(nodeOrders))
 		for _, ele := range nodeOrders {
-			// if order.Type == Sell {
-			// 	if ele.Price.Cmp(order.Price) == -1 {
-			// 		noMoreOrders = true
-			// 		break
-			// 	}
-			// } else {
-			// 	if ele.Price.Cmp(order.Price) == 1 {
-			// 		noMoreOrders = true
-			// 		break
-			// 	}
-			// }
+			if order.Type == Sell {
+				if ele.Price.Cmp(order.Price) == -1 {
+					noMoreOrders = true
+					break
+				}
+			} else {
+				if ele.Price.Cmp(order.Price) == 1 {
+					noMoreOrders = true
+					break
+				}
+			}
 
 			// countAdd += ele.Amount
-			// fmt.Println(ele.Price, ele.Amount, order.Amount, ele.Amount.Cmp(order.Amount))
+			// fmt.Println(ele.Price, order.Price, ele.Amount, order.Amount, ele.Amount.Cmp(order.Amount))
 			if ele.Amount.Cmp(order.Amount) == 1 {
 				nodeData.updateVolume(order.Amount.Neg())
 				// trades = append(trades, Trade{BuyOrderID: ele.ID, SellOrderID: order.ID, Amount: order.Amount, Price: ele.Price})
@@ -139,20 +161,19 @@ func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTre
 				ele.Amount = amount
 
 				partialOrder = NewOrder(ele.ID, ele.Type, ele.Amount, ele.Price)
-				ordersProcessed = append(ordersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, decimalZero))
+				ordersProcessed = append(ordersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, order.Price))
 
 				maxNode.SetData(nodeData)
 
 				order.Amount, _ = util.NewDecimalFromString("0.0")
 				noMoreOrders = true
 				break
-			}
-			if ele.Amount.Cmp(order.Amount) == 0 {
+			} else if ele.Amount.Cmp(order.Amount) == 0 {
 				nodeData.updateVolume(order.Amount.Neg())
 
 				ordersProcessed = append(ordersProcessed, NewOrder(ele.ID, ele.Type, ele.Amount, ele.Price))
-				ordersProcessed = append(ordersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, decimalZero))
-
+				ordersProcessed = append(ordersProcessed, NewOrder(order.ID, order.Type, orderOriginalAmount, order.Price))
+				partialOrder = nil
 				countMatch++
 				// trades = append(trades, Trade{BuyOrderID: ele.ID, SellOrderID: order.ID, Amount: order.Amount, Price: ele.Price})
 
@@ -167,6 +188,10 @@ func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTre
 				break
 			} else {
 				countMatch++
+
+				amount := order.Amount.Sub(ele.Amount)
+
+				partialOrder = NewOrder(order.ID, order.Type, amount, order.Price)
 
 				ordersProcessed = append(ordersProcessed, NewOrder(ele.ID, ele.Type, ele.Amount, ele.Price))
 
@@ -191,5 +216,7 @@ func (ob *OrderBook) processLimitMarket(order *Order, tree *binarytree.BinaryTre
 		maxNode.SetData(nodeData)
 	}
 	// return trades, noMoreOrders, ordersProcessed, partialOrder
+	// fmt.Printf("partialOrder: %#v\n", partialOrder)
+	// fmt.Printf("order.Amount: %#v\n", order.Amount.String())
 	return noMoreOrders, ordersProcessed, partialOrder
 }
