+++ internal/app/server/engine.go
@@ -6,27 +6,27 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/Pantelwar/matching-engine/engine"
-	engineGrpc "github.com/Pantelwar/matching-engine/engineGrpc"
-	"github.com/Pantelwar/matching-engine/util"
+	engine2 "github.com/Pantelwar/matching-engine/internal/app/engine"
+	engine3 "github.com/Pantelwar/matching-engine/internal/app/engineGrpc"
+	"github.com/Pantelwar/matching-engine/internal/app/util"
 )
 
 // Engine ...
 type Engine struct {
-	book map[string]*engine.OrderBook
+	book map[string]*engine2.OrderBook
 }
 
 // NewEngine returns Engine object
 func NewEngine() *Engine {
-	return &Engine{book: map[string]*engine.OrderBook{}}
+	return &Engine{book: map[string]*engine2.OrderBook{}}
 }
 
 // Process implements EngineServer interface
-func (e *Engine) Process(ctx context.Context, req *engineGrpc.Order) (*engineGrpc.OutputOrders, error) {
+func (e *Engine) Process(ctx context.Context, req *engine3.Order) (*engine3.OutputOrders, error) {
 	bigZero, _ := util.NewDecimalFromString("0.0")
 	orderString := fmt.Sprintf("{\"id\":\"%s\", \"type\": \"%s\", \"amount\": \"%s\", \"price\": \"%s\" }", req.GetID(), req.GetType(), req.GetAmount(), req.GetPrice())
 
-	var order engine.Order
+	var order engine2.Order
 	// decode the message
 	fmt.Println("Orderstring =: ", orderString)
 	err := order.FromJSON([]byte(orderString))
@@ -45,11 +45,11 @@ func (e *Engine) Process(ctx context.Context, req *engineGrpc.Order) (*engineGrp
 		return nil, errors.New("Invalid pair")
 	}
 
-	var pairBook *engine.OrderBook
+	var pairBook *engine2.OrderBook
 	if val, ok := e.book[req.GetPair()]; ok {
 		pairBook = val
 	} else {
-		pairBook = engine.NewOrderBook()
+		pairBook = engine2.NewOrderBook()
 		e.book[req.GetPair()] = pairBook
 	}
 
@@ -74,14 +74,14 @@ func (e *Engine) Process(ctx context.Context, req *engineGrpc.Order) (*engineGrp
 			fmt.Println("partialOrderString Marshal error", err)
 			return nil, err
 		}
-		return &engineGrpc.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: string(partialOrderString)}, nil
+		return &engine3.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: string(partialOrderString)}, nil
 	}
-	return &engineGrpc.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: "null"}, nil
+	return &engine3.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: "null"}, nil
 }
 
 // Cancel implements EngineServer interface
-func (e *Engine) Cancel(ctx context.Context, req *engineGrpc.Order) (*engineGrpc.Order, error) {
-	order := &engine.Order{ID: req.GetID()}
+func (e *Engine) Cancel(ctx context.Context, req *engine3.Order) (*engine3.Order, error) {
+	order := &engine2.Order{ID: req.GetID()}
 
 	if order.ID == "" {
 		fmt.Println("Invalid JSON")
@@ -93,11 +93,11 @@ func (e *Engine) Cancel(ctx context.Context, req *engineGrpc.Order) (*engineGrpc
 		return nil, errors.New("Invalid pair")
 	}
 
-	var pairBook *engine.OrderBook
+	var pairBook *engine2.OrderBook
 	if val, ok := e.book[req.GetPair()]; ok {
 		pairBook = val
 	} else {
-		pairBook = engine.NewOrderBook()
+		pairBook = engine2.NewOrderBook()
 		e.book[req.GetPair()] = pairBook
 	}
 
@@ -110,22 +110,22 @@ func (e *Engine) Cancel(ctx context.Context, req *engineGrpc.Order) (*engineGrpc
 		return nil, errors.New("NoOrderPresent")
 	}
 
-	orderEngine := &engineGrpc.Order{}
+	orderEngine := &engine3.Order{}
 
 	orderEngine.ID = order.ID
 	orderEngine.Amount = order.Amount.String()
 	orderEngine.Price = order.Price.String()
-	orderEngine.Type = engineGrpc.Side(engineGrpc.Side_value[order.Type.String()])
+	orderEngine.Type = engine3.Side(engine3.Side_value[order.Type.String()])
 
 	return orderEngine, nil
 }
 
 // ProcessMarket implements EngineServer interface
-func (e *Engine) ProcessMarket(ctx context.Context, req *engineGrpc.Order) (*engineGrpc.OutputOrders, error) {
+func (e *Engine) ProcessMarket(ctx context.Context, req *engine3.Order) (*engine3.OutputOrders, error) {
 	bigZero, _ := util.NewDecimalFromString("0.0")
 	orderString := fmt.Sprintf("{\"id\":\"%s\", \"type\": \"%s\", \"amount\": \"%s\", \"price\": \"%s\" }", req.GetID(), req.GetType(), req.GetAmount(), req.GetPrice())
 
-	var order engine.Order
+	var order engine2.Order
 	// decode the message
 	// fmt.Println("Orderstring =: ", orderString)
 	err := order.FromJSON([]byte(orderString))
@@ -144,11 +144,11 @@ func (e *Engine) ProcessMarket(ctx context.Context, req *engineGrpc.Order) (*eng
 		return nil, errors.New("Invalid pair")
 	}
 
-	var pairBook *engine.OrderBook
+	var pairBook *engine2.OrderBook
 	if val, ok := e.book[req.GetPair()]; ok {
 		pairBook = val
 	} else {
-		pairBook = engine.NewOrderBook()
+		pairBook = engine2.NewOrderBook()
 		e.book[req.GetPair()] = pairBook
 	}
 
@@ -168,19 +168,19 @@ func (e *Engine) ProcessMarket(ctx context.Context, req *engineGrpc.Order) (*eng
 	if partialOrder != nil {
 		var partialOrderString []byte
 		partialOrderString, err = json.Marshal(partialOrder)
-		return &engineGrpc.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: string(partialOrderString)}, nil
+		return &engine3.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: string(partialOrderString)}, nil
 	}
-	return &engineGrpc.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: "null"}, nil
+	return &engine3.OutputOrders{OrdersProcessed: string(ordersProcessedString), PartialOrder: "null"}, nil
 }
 
 // FetchBook implements EngineServer interface
-func (e *Engine) FetchBook(ctx context.Context, req *engineGrpc.BookInput) (*engineGrpc.BookOutput, error) {
+func (e *Engine) FetchBook(ctx context.Context, req *engine3.BookInput) (*engine3.BookOutput, error) {
 	if req.GetPair() == "" {
 		fmt.Println("Invalid pair")
 		return nil, errors.New("Invalid pair")
 	}
 
-	var pairBook *engine.OrderBook
+	var pairBook *engine2.OrderBook
 	if val, ok := e.book[req.GetPair()]; ok {
 		pairBook = val
 	} else {
@@ -190,39 +190,39 @@ func (e *Engine) FetchBook(ctx context.Context, req *engineGrpc.BookInput) (*eng
 	fmt.Println(pairBook)
 	book := pairBook.GetOrders(req.GetLimit())
 
-	result := &engineGrpc.BookOutput{Buys: []*engineGrpc.BookArray{}, Sells: []*engineGrpc.BookArray{}}
+	result := &engine3.BookOutput{Buys: []*engine3.BookArray{}, Sells: []*engine3.BookArray{}}
 
 	for _, buy := range book.Buys {
-		arr := &engineGrpc.BookArray{PriceAmount: []string{}}
+		arr := &engine3.BookArray{PriceAmount: []string{}}
 
 		bodyBytes, err := json.Marshal(buy)
 		if err != nil {
 			fmt.Println("1", err)
-			return &engineGrpc.BookOutput{Buys: []*engineGrpc.BookArray{}, Sells: []*engineGrpc.BookArray{}}, nil
+			return &engine3.BookOutput{Buys: []*engine3.BookArray{}, Sells: []*engine3.BookArray{}}, nil
 		}
 
 		err = json.Unmarshal(bodyBytes, &arr.PriceAmount)
 		if err != nil {
 			fmt.Println("2", err)
-			return &engineGrpc.BookOutput{Buys: []*engineGrpc.BookArray{}, Sells: []*engineGrpc.BookArray{}}, nil
+			return &engine3.BookOutput{Buys: []*engine3.BookArray{}, Sells: []*engine3.BookArray{}}, nil
 		}
 
 		result.Buys = append(result.Buys, arr)
 	}
 
 	for _, sell := range book.Sells {
-		arr := &engineGrpc.BookArray{PriceAmount: []string{}}
+		arr := &engine3.BookArray{PriceAmount: []string{}}
 
 		bodyBytes, err := json.Marshal(sell)
 		if err != nil {
 			fmt.Println("json.Marshal Error", err)
-			return &engineGrpc.BookOutput{Buys: []*engineGrpc.BookArray{}, Sells: []*engineGrpc.BookArray{}}, nil
+			return &engine3.BookOutput{Buys: []*engine3.BookArray{}, Sells: []*engine3.BookArray{}}, nil
 		}
 
 		err = json.Unmarshal(bodyBytes, &arr.PriceAmount)
 		if err != nil {
 			fmt.Println("json.Unmarshal Error", err)
-			return &engineGrpc.BookOutput{Buys: []*engineGrpc.BookArray{}, Sells: []*engineGrpc.BookArray{}}, nil
+			return &engine3.BookOutput{Buys: []*engine3.BookArray{}, Sells: []*engine3.BookArray{}}, nil
 		}
 
 		result.Sells = append(result.Sells, arr)
